Add DateTimeModel with SplitDateAndTime Spec driven test cases
