+++ JavaScript/src/dateTime/dateTimeRecognizer.ts
@@ -2,9 +2,13 @@ import { IModel } from "../models";
 import { IDateTimeModel, DateTimeModel } from "./models";
 import { Recognizer } from "../recognizer";
 import { Culture } from "../culture";
-import { BaseMergedParser, BaseMergedExtractor, DateTimeOptions } from "./baseMerged";
+import { BaseMergedParser, BaseMergedExtractor } from "./baseMerged";
 import { EnglishCommonDateTimeParserConfiguration } from "./english/baseConfiguration";
 import { EnglishMergedExtractorConfiguration, EnglishMergedParserConfiguration } from "./english/mergedConfiguration";
+
+export enum DateTimeOptions {
+    None = 0, SkipFromToMerge = 1, SplitDateAndTime = 2
+}
 
 export default class DateTimeRecognizer extends Recognizer {
     static readonly instance: DateTimeRecognizer = new DateTimeRecognizer(DateTimeOptions.None);
