+++ JavaScript/test/mergedExtractor-english.test.js
@@ -1,7 +1,7 @@
 var describe = require('ava-spec').describe;
 var EnglishMergedExtractorConfiguration = require('../compiled/dateTime/english/mergedConfiguration').EnglishMergedExtractorConfiguration;
 var BaseMergedExtractor = require('../compiled/dateTime/baseMerged').BaseMergedExtractor;
-var DateTimeOptions = require('../compiled/dateTime/baseMerged').DateTimeOptions;
+var DateTimeOptions = require('../compiled/dateTime/dateTimeRecognizer').DateTimeOptions;
 var Constants = require('../compiled/dateTime/constants').Constants;
 
 describe('Merged Extractor', it => {
